Move field solver into a package
