+++ src/ef/simulation.py
@@ -1,7 +1,7 @@
 import h5py
 import numpy as np
 
-from ef.field_solver import FieldSolver
+from ef.field.solvers.field_solver import FieldSolver
 from ef.util.serializable_h5 import SerializableH5
